# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AeroFS API",
    "description": "Remote file manager REST API. All endpoints require session authentication except public share endpoints.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/trash": {
      "get": {
        "tags": [
          "trash"
        ],
        "summary": "List all items in trash",
        "operationId": "list_trash",
        "responses": {
          "200": {
            "description": "List of trash items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrashItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/trash/empty": {
      "delete": {
        "tags": [
          "trash"
        ],
        "summary": "Empty entire trash",
        "operationId": "empty_trash",
        "responses": {
          "200": {
            "description": "Trash emptied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyTrashResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/trash/move": {
      "post": {
        "tags": [
          "trash"
        ],
        "summary": "Move one or more items to trash (soft delete)",
        "operationId": "move_to_trash",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToTrashRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Items moved to trash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveToTrashResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/trash/restore/{id}": {
      "post": {
        "tags": [
          "trash"
        ],
        "summary": "Restore an item from trash back to its original location",
        "operationId": "restore_trash_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Trash item ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item restored from trash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrashActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/trash/{id}": {
      "delete": {
        "tags": [
          "trash"
        ],
        "summary": "Delete an item permanently from trash",
        "operationId": "delete_trash_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Trash item ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item permanently deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrashActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyTrashResponse": {
        "type": "object",
        "required": [
          "success",
          "deleted_count",
          "message"
        ],
        "properties": {
          "deleted_count": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ErrorCategory": {
        "type": "string",
        "enum": [
          "validation",
          "authentication",
          "authorization",
          "not_found",
          "conflict",
          "transfer_cancelled",
          "payload_too_large",
          "insufficient_storage",
          "rate_limited",
          "timeout",
          "provider",
          "io",
          "security",
          "internal"
        ]
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "category",
          "retryable",
          "message"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategory"
          },
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "user_action": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "MoveToTrashRequest": {
        "type": "object",
        "required": [
          "connection_id",
          "paths"
        ],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MoveToTrashResponse": {
        "type": "object",
        "required": [
          "success",
          "moved_count",
          "moved_items",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "moved_count": {
            "type": "integer",
            "minimum": 0
          },
          "moved_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovedTrashItem"
            }
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "MovedTrashItem": {
        "type": "object",
        "required": [
          "id",
          "original_path"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "original_path": {
            "type": "string"
          }
        }
      },
...
      "TrashActionResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "TrashItem": {
        "type": "object",
        "required": [
          "id",
          "connection_id",
          "original_path",
          "item_name",
          "is_directory",
          "deleted_at"
        ],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "deleted_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "is_directory": {
            "type": "boolean"
          },
          "item_name": {
            "type": "string"
          },
          "original_path": {
            "type": "string"
          },
          "size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aerofs_session"
      }
    }
  },
...
}